+++ nest-note.xcodeproj/xcuserdata/cswapp.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -39,22 +39,6 @@
       <BreakpointProxy
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
-            uuid = "126827D2-002A-4941-B0F8-039B55745946"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "nest-note/Scenes/Places/PlaceDetailViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "305"
-            endingLineNumber = "305"
-            landmarkName = "saveButtonTapped()"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
             uuid = "DF29AEE1-22EE-460F-AC61-8076F775C71F"
             shouldBeEnabled = "Yes"
             ignoreCount = "0"
