+++ nest-note/Extensions/Notification+Extensions.swift
@@ -4,4 +4,5 @@ extension Notification.Name {
     static let userInformationUpdated = Notification.Name("userInformationUpdated")
     static let setupStepDidComplete = Notification.Name("setupStepDidComplete")
     static let entryDidSave = Notification.Name("entryDidSave")
+    static let placeDidSave = Notification.Name("placeDidSave")
 }
