+++ nest-note/Scenes/Onboarding/OBCreateNestViewController.swift
@@ -1,3 +1,12 @@
+//
+//  OBCreateNestViewController.swift
+//  nest-note
+//
+//  Created by Colton Swapp on 7/8/25.
+//
+
+import UIKit
+
 class OBCreateNestViewController: NNOnboardingViewController {
     private let nestNameField: NNTextField = {
         let field = NNTextField()
@@ -6,10 +15,25 @@ class OBCreateNestViewController: NNOnboardingViewController {
         return field
     }()
     
+    private let nameFootnoteLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = "'Smith Nest' or 'Anderson Nest' is recommended."
+        label.textColor = .secondaryLabel
+        label.font = .preferredFont(forTextStyle: .caption1)
+        label.numberOfLines = 0
+        label.textAlignment = .center
+        return label
+    }()
+    
     private let addressField: NNTextField = {
         let field = NNTextField()
         field.translatesAutoresizingMaskIntoConstraints = false
         field.placeholder = "321 Eagle Nest Ct, Birdsville CA"
+        field.textContentType = .fullStreetAddress
+        field.keyboardType = .default
+        field.autocapitalizationType = .words
+        field.autocorrectionType = .no
         return field
     }()
     
@@ -79,6 +103,7 @@ class OBCreateNestViewController: NNOnboardingViewController {
     // MARK: - Setup
     override func setupContent() {
         view.addSubview(nestNameField)
+        view.addSubview(nameFootnoteLabel)
         view.addSubview(addressField)
         view.addSubview(addressFootnoteLabel)
         
@@ -88,7 +113,11 @@ class OBCreateNestViewController: NNOnboardingViewController {
             nestNameField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
             nestNameField.heightAnchor.constraint(equalToConstant: 56),
             
-            addressField.topAnchor.constraint(equalTo: nestNameField.bottomAnchor, constant: 16),
+            nameFootnoteLabel.topAnchor.constraint(equalTo: nestNameField.bottomAnchor, constant: 8),
+            nameFootnoteLabel.leadingAnchor.constraint(equalTo: nestNameField.leadingAnchor, constant: 8),
+            nameFootnoteLabel.trailingAnchor.constraint(equalTo: nestNameField.trailingAnchor, constant: -8),
+            
+            addressField.topAnchor.constraint(equalTo: nameFootnoteLabel.bottomAnchor, constant: 16),
             addressField.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
             addressField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
             addressField.heightAnchor.constraint(equalToConstant: 56),
